pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaz-video-dub"
version = "0.1.0"
description = "English to Kazakh video dubbing pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Orenbek" }]
dependencies = [
  "pydantic>=2.7,<3",
  "typer>=0.12,<1",
  "rich>=13,<14",
  "PyYAML>=6,<7",
  "srt>=3.5,<4",
]

[project.scripts]
video-dub = "video_dub.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
